// src/graphql-types.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]?: Maybe<T[SubKey]>;
};
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]: Maybe<T[SubKey]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  /** The `ID` scalar type represents a unique identifier, often used to refetch an object or as key for a cache. The ID type appears in a JSON response as a String; however, it is not intended to be human-readable. When expected as an input type, any string (such as `"4"`) or integer (such as `4`) input value will be accepted as an ID. */
  ID: string;
  /** The `String` scalar type represents textual data, represented as UTF-8 character sequences. The String type is most often used by GraphQL to represent free-form human-readable text. */
  String: string;
  /** The `Boolean` scalar type represents `true` or `false`. */
  Boolean: boolean;
  /** The `Int` scalar type represents non-fractional signed whole numeric values. Int can represent values between -(2^31) and 2^31 - 1. */
  Int: number;
  /** The `Float` scalar type represents signed double-precision fractional values as specified by [IEEE 754](https://en.wikipedia.org/wiki/IEEE_floating_point). */
  Float: number;
  /** A date string, such as 2007-12-03, compliant with the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  Date: any;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
};
// ...
export type AirtableData = {
  Beneficiary?: Maybe<Scalars["String"]>;
  Grant_origin?: Maybe<Scalars["String"]>;
  Project_name?: Maybe<Scalars["String"]>;
  Category?: Maybe<Scalars["String"]>;
  Status?: Maybe<Scalars["String"]>;
  Project_Lead?: Maybe<Scalars["String"]>;
  origin_link?: Maybe<Scalars["String"]>;
  Currency?: Maybe<Scalars["String"]>;
  Fund?: Maybe<Scalars["String"]>;
  Sub_beneficiary?: Maybe<Scalars["String"]>;
  USD_equivalent?: Maybe<Scalars["Int"]>;
  Last_Modified?: Maybe<Scalars["Date"]>;
  Funds_requested?: Maybe<Scalars["Int"]>;
  Started_on?: Maybe<Scalars["Date"]>;
  Completed_on?: Maybe<Scalars["Date"]>;
  Description?: Maybe<Scalars["String"]>;
  Status_title?: Maybe<Scalars["String"]>;
  Status_description?: Maybe<Scalars["String"]>;
  Status_desc_link1?: Maybe<Scalars["String"]>;
};